requirements.txt
Flask==2.0.3
Keras==2.3.1
gunicorn==20.1.0
opencv-python==4.5.5.62
Pillow==9.0.0
pyttsx3==2.90
scikit-learn==1.0.2
scipy==1.4.1
tensorflow-gpu==2.2.0
numpy==1.19.5